Добавлен диспенсер надувных дверей и стен для киборгов, добавлены новые реагенты в гипоспрей киборга. Возможность брать манипулятором части мехов, аптечки и части для создания ботов.

+++ code/game/objects/structures/inflatable.dm
@@ -1,17 +1,31 @@
 /obj/item/inflatable
+	name = "inflatable"
+	w_class = 2
+	icon = 'icons/obj/inflatable.dmi'
+	var/deploy_path = null
+
+/obj/item/inflatable/attack_self(mob/user)
+	if(!deploy_path)
+		return
+	playsound(loc, 'sound/items/zip.ogg', 75, 1)
+	user << "<span class='notice'>You inflate \the [src].</span>"
+	var/obj/structure/inflatable/R = new deploy_path(user.loc)
+	src.transfer_fingerprints_to(R)
+	R.add_fingerprint(user)
+	qdel(src)
+
+
+/obj/item/inflatable/wall
 	name = "inflatable wall"
 	desc = "A folded membrane which rapidly expands into a large cubical shape on activation."
-	icon = 'icons/obj/inflatable.dmi'
 	icon_state = "folded_wall"
-	w_class = 3
+	deploy_path = /obj/structure/inflatable/wall
 
-	attack_self(mob/user)
-		playsound(loc, 'sound/items/zip.ogg', 75, 1)
-		user << "\blue You inflate [src]."
-		var/obj/structure/inflatable/R = new /obj/structure/inflatable(user.loc)
-		src.transfer_fingerprints_to(R)
-		R.add_fingerprint(user)
-		qdel(src)
+/obj/item/inflatable/door/
+	name = "inflatable door"
+	desc = "A folded membrane which rapidly expands into a simple door on activation."
+	icon_state = "folded_door"
+	deploy_path = /obj/structure/inflatable/door
 
 /obj/structure/inflatable
 	name = "inflatable wall"
@@ -23,7 +37,12 @@
 	icon = 'icons/obj/inflatable.dmi'
 	icon_state = "wall"
 
+	var/undeploy_path = null
 	var/health = 50.0
+
+/obj/structure/inflatable/wall
+	name = "inflatable wall"
+	undeploy_path = /obj/item/inflatable/wall
 
 
 /obj/structure/inflatable/New(location)
@@ -40,7 +59,7 @@
 /obj/structure/inflatable/bullet_act(var/obj/item/projectile/Proj)
 	if(!(Proj.damage_type == BRUTE || Proj.damage_type == BURN))
 		return
-	
+
 	health -= Proj.damage
 	..()
 	if(health <= 0)
@@ -67,11 +86,11 @@
 	deflate(1)
 
 /obj/structure/inflatable/attack_hand(mob/user as mob)
-		add_fingerprint(user)
-		return
+	add_fingerprint(user)
+	return
 
 /obj/structure/inflatable/attackby(obj/item/weapon/W as obj, mob/user as mob)
-	if(!istype(W)) return
+	if(!istype(W) || istype(W, /obj/item/weapon/inflatable_dispenser)) return
 
 	if (can_puncture(W))
 		visible_message("\red <b>[user] pierces [src] with [W]!</b>")
@@ -96,10 +115,11 @@
 		src.transfer_fingerprints_to(R)
 		qdel(src)
 	else
-		//user << "\blue You slowly deflate the inflatable wall."
-		visible_message("[src] slowly deflates.")
+		if(!undeploy_path)
+			return
+		visible_message("\The [src] slowly deflates.")
 		spawn(50)
-			var/obj/item/inflatable/R = new /obj/item/inflatable(loc)
+			var/obj/item/inflatable/R = new undeploy_path(src.loc)
 			src.transfer_fingerprints_to(R)
 			qdel(src)
 
@@ -124,26 +144,13 @@
 		user.visible_message("<span class='danger'>[user] [attack_verb] at [src]!</span>")
 	return 1
 
-/obj/item/inflatable/door/
-	name = "inflatable door"
-	desc = "A folded membrane which rapidly expands into a simple door on activation."
-	icon = 'icons/obj/inflatable.dmi'
-	icon_state = "folded_door"
-
-	attack_self(mob/user)
-		playsound(loc, 'sound/items/zip.ogg', 75, 1)
-		user << "\blue You inflate [src]."
-		var/obj/structure/inflatable/door/R = new /obj/structure/inflatable/door(user.loc)
-		src.transfer_fingerprints_to(R)
-		R.add_fingerprint(user)
-		qdel(src)
-
 /obj/structure/inflatable/door //Based on mineral door code
 	name = "inflatable door"
 	density = 1
 	anchored = 1
 	opacity = 0
 
+	undeploy_path = /obj/item/inflatable/door
 	icon = 'icons/obj/inflatable.dmi'
 	icon_state = "door_closed"
 
@@ -171,7 +178,6 @@
 	if(isSwitchingStates) return
 	if(ismob(user))
 		var/mob/M = user
-		if(world.time - user.last_bumped <= 60) return //NOTE do we really need that?
 		if(M.client)
 			if(iscarbon(M))
 				var/mob/living/carbon/C = M
@@ -262,7 +268,7 @@
 		new /obj/item/inflatable/door(src)
 		new /obj/item/inflatable/door(src)
 		new /obj/item/inflatable/door(src)
-		new /obj/item/inflatable(src)
-		new /obj/item/inflatable(src)
-		new /obj/item/inflatable(src)
-		new /obj/item/inflatable(src)
+		new /obj/item/inflatable/wall(src)
+		new /obj/item/inflatable/wall(src)
+		new /obj/item/inflatable/wall(src)
+		new /obj/item/inflatable/wall(src)
